Add some more utilities to assertions
